% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county-population-data.R
\docType{data}
\name{county_population}
\alias{county_population}
\title{County wise population with FIPS}
\format{
An dataframe with 32 columns and 3221 observations.
}
\usage{
county_population
}
\description{
The dataset contains population for all counties along Unites States
from 1990 to 2019 with FIPS code. It is used for evaluation of density metrics.
It has been built using censusapi Population estimates pep/int_population for 2000-2009,
American Community Survey acs/acs5 for 2010-2019 and Census beaureau population data for
1990-1999.
}
\keyword{datasets}
