% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\name{get_cr}
\alias{get_cr}
\title{Concentration Ratio}
\usage{
get_cr(df, geo, subsector, resource, number)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(concentration ratio) grouped by geographical level and sub sector.}

\item{geo}{A character string indicating the column name for georaphical area unique code.}

\item{subsector}{A character string indicating the column name for sub sectors.}

\item{resource}{A character string indicating the column name for resource based on which cr is calculated. Example: Revenue, assets etc.}

\item{number}{A numeric value that is used to calculate concentration ratio for top that number of firms. Example, 4 if concentration ratiofor top 4 firms.}
}
\value{
A new dataframe with the Concentration Ratio of n(number) largest organizations ('cr'),
grouped by geographical level and sub sectors *Note: For the markets where total number of firms is less than n(umber),
the concentration ratio is assigned 100\%.
}
\description{
Calculate the Concentration Ratio (CR) of top n(number) largest organizations and return a dataframe.
}
\details{
Concentration Ratio is defined as the summation of the market shares of the largest n firms.
It is calculated using the summation of revenue of the largest n nonprofits divided by the summation of
revenue of all nonprofits in a market. The formula is \eqn{\frac{\sigma{revnue of top-n firms}}{Total resource of that subsector}}
}
\examples{
head( dat.cr )

}
