% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr4.R
\name{get_cr2}
\alias{get_cr2}
\title{Concentration Ratio -2}
\usage{
get_cr2(df, MSA_NECH, NTMAJ12, TOTREV)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(cr2) grouped by MSA_NECH and sub sector.,}

\item{MSA_NECH}{A character string indicating the column name for Metropolitan Statistical Area unique code.}

\item{NTMAJ12}{A character string indicating the column name for sub sectors.}

\item{TOTREV}{A character string indicating the column name for total revenue.}
}
\value{
A new dataframe with the Concentration Ratio of 4 largest organizations ('cr4'),
grouped by MSA_NECH.
}
\description{
Calculate the Concentration Ratio (CR) of top 2 largest organizations and return a dataframe.
}
\details{
CR2 is defined as the summation of the market shares of the largest two firms.
It is calculated using the summation of revenue of the largest two nonprofits divided by the summation of
revenue of all nonprofits in a market.
}
\examples{
x1 <- floor(runif(5, min=10000, max=99999))
head( a )

}
