% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_density}
\alias{get_density}
\title{Density metric}
\usage{
get_density(df, geo, subsector, fips, year)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(density) grouped by gepgraphical level and sub sector.}

\item{geo}{A character string indicating the column name for geographical area unique code.}

\item{subsector}{A character string indicating the column name for sub sectors.}

\item{fips}{A character string indicating the column name for FIPS code of the geographical area}

\item{year}{A numeric parameter indicating the year to be used for population estimate. Example: 1999}
}
\value{
A new dataframe with the density of nonprofits per 10 and 1000 population,
grouped by geographical level and sub sectors.
}
\description{
Calculate the nonprofit density metric and return a dataframe.
}
\details{
The density metric is the number of nonprofits per-capita in a market. In this function, it is measured
as the number of organizations per 1000 population and per 10,000 population.
}
\examples{
head( dat.den.2009 )

}
