% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_density_commercial}
\alias{get_density_commercial}
\title{Density commercial}
\usage{
get_density_commercial(df, geo, subsector, resource, list.of.cols)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(density commercial) grouped by geographical level and sub sector.}

\item{geo}{A character string indicating the column name for geographical area unique code.}

\item{subsector}{A character string indicating the column name for sub sectors.}

\item{resource}{A character string indicating the column name for resource based on which density is evaluated}

\item{list.of.cols}{A vector containing list of column names that have the commercial metrics for the resource}
}
\value{
A new dataframe with the commercial density of nonprofits summarized in the subsector and geo level.
}
\description{
Calculate the commercial nonprofit density  a resource  and return a dataframe.
}
\details{
The density metric is the number of nonprofits per-capita in a market. In this function, it is measured
as the number of organizations per 1000 population and per 10,000 population.
}
\examples{
dat.den.comm <- get_density_commercial( df=nonprofit_sample,'MSA_NECH','NTMAJ12', 'TOTREV', c('PROGREV', 'INVINC'))
head( dat.den.comm )

}
