% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_density_small}
\alias{get_density_small}
\title{Density small}
\usage{
get_density_small(df, geo, subsector, fips, year, resource, smaller.than)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(density) grouped by gepgraphical level and sub sector.}

\item{geo}{A character string indicating the column name for geographical area unique code.}

\item{subsector}{A character string indicating the column name for sub sectors.}

\item{fips}{A character string indicating the column name for FIPS code of the geographical area}

\item{year}{A numeric parameter indicating the year to be used for population estimate. Example: 1999}

\item{resource}{A character string indicating the column name for resource based on which density is evaluated}

\item{smaller.than}{A numeric value indicating that it is the upper limit of the resource for density evaluation. All organizations having resource value less than that would be counted for evaluation.}
}
\value{
A new dataframe with the density of nonprofits per 10 and 1000 population, having resource smaller than the given number
grouped by geographical level and sub sectors.
}
\description{
Calculate the nonprofit density metric having resource smaller than a numeric value and return a dataframe.
}
\details{
The density metric is the number of nonprofits per-capita in a market. In this function, it is measured
as the number of organizations per 1000 population and per 10,000 population.
}
\examples{
dat.den.2009 <- get_density_small( df=nonprofit_sample,'MSA_NECH','NTMAJ12','FIPS', 2009, 'TOTREV', 100000)
head( dat.den.2009 )

}
