% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundraising-efficiency.R
\name{get_fundraising_efficiency}
\alias{get_fundraising_efficiency}
\title{Fundraising efficiency}
\usage{
get_fundraising_efficiency(df, geo, subsector, resource, list.of.cols)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(fundraising efficiency) grouped by geographical level and sub sector.}

\item{geo}{A character string indicating the column name for geographical area unique code.}

\item{subsector}{A character string indicating the column name for sub sectors.}

\item{resource}{A character string indicating the column name for resource based on which fundraising efficiency is evaluated}

\item{list.of.cols}{A vector containing list of column names that have the fundraising fees for the nonprofits.}
}
\value{
A new dataframe with the fundraising efficiency of nonprofits, grouped by geographical level and sub sectors.
}
\description{
Calculate the fundraising efficiency of nonprofits in a geographical level  and return a dataframe.
}
\details{
Fundraising efficiency is the measure of the fundraising competition in a market. It represents total private donations raised relative
to expenses on fundraising activities and indicates the amount of private donations raised for each dollar of fundraising expense incurred.
Lower ratio indicates greater fundraising competition in a market.Fundraising efficiency be calculated as follows: Private donations divided by fundraising expenses
}
\examples{
data(nonprofit_sample)
dat.fe <- get_density_commercial( df=nonprofit_sample,'MSA_NECH','NTMAJ12', 'CONT', c('FUNDFEES', 'SOLICIT'))
head( dat.fe)

}
