% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwoka-index.R
\name{get_kwoka_index}
\alias{get_kwoka_index}
\title{Kwoka Index}
\usage{
get_kwoka_index(df, MSA_NECH, NTMAJ12, TOTREV)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(kwoka index) grouped by MSA_NECH and sub sector.}

\item{MSA_NECH}{A character string indicating the column name for Metropolitan Statistical Area unique code.}

\item{NTMAJ12}{A character string indicating the column name for sub sectors.}

\item{TOTREV}{A character string indicating the column name for total revenue.}
}
\value{
A new dataframe with the Kwoka Index (\code{kindex}),
grouped by MSA_NECH and NTMAJ12(12 sub sectors).
}
\description{
Calculate the Kwoka Index (kindex) and return a dataframe.
}
\details{
The Kwoka Index(kindex) is a measure of market concentration that is used to understand market competitiveness,
competitive behavior and performance in markets.
}
\examples{
x1 <- floor(runif(5, min=10000, max=99999))
head( a )

}
