% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonprofit-sample-data.R
\docType{data}
\name{nonprofit_sample}
\alias{nonprofit_sample}
\title{Sample data from NCCS data archive}
\format{
An dataframe with 152 columns and 75814 observations.
}
\source{
\url{https://nccs-data.urban.org/}
}
\usage{
nonprofit_sample
}
\description{
The dataset contains sample data from NCCS data archives core Public charities
}
\keyword{datasets}
