% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwoka-index.R
\name{get_kwoka_index}
\alias{get_kwoka_index}
\title{Kwoka Index}
\usage{
get_kwoka_index(df, geo, subsector, resource)
}
\arguments{
\item{df}{A dataframe containing the required fields for computing the metric(kwoka index) grouped by geographical level and sub sector.}

\item{geo}{A character string indicating the column name for georaphical area unique code.}

\item{subsector}{A character string indicating the column name for sub sectors.}

\item{resource}{A character string indicating the column name for resource based on which cr4 is calculated. Example: Revenue, assets etc.}
}
\value{
A new dataframe with the Kwoka Index (\code{kindex}),
grouped by geographical level and sub sectors.
}
\description{
Calculate the Kwoka Index (kindex) and return a dataframe.
}
\details{
The Kwoka Index(kindex) is a measure of market concentration that is used to understand market competitiveness,
competitive behavior and performance in markets.
}
\examples{
data(nonprofit_sample)
head( dat.kindex )

}
